New composer API, bump WIP version.
