+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "aery"
-version = "0.2.0"
+version = "0.3.0"
 edition = "2021"
 authors = ["iiYese iiyese@outlook.com"]
 repository = "https://github.com/iiYese/aery"
