+++ src/relation.rs
@@ -99,7 +99,7 @@ pub enum CleanupPolicy {
     /// will delete themselves. This is effectively reference counting.
     Counted,
     /// When targets of recursively cleaning relations are deleted they also delete all their
-    /// hosts. Unsetting a recursively cleaning relation is the same as despawning the host.
+    /// hosts. Untargetting a recursively cleaning relation is the same as despawning the host.
     Recursive,
     /// Total performs both counted and recursive cleanup.
     Total,
@@ -107,6 +107,8 @@ pub enum CleanupPolicy {
 
 pub trait ZstOrPanic: Sized {
     const ZST_OR_PANIC: () = {
+        // TODO: Make diagnostic friendlier when `std::any::type_name` becomes const
+        // TODO: Use actual type level mechanism and remove hack when possible in stable
         if std::mem::size_of::<Self>() != 0 {
             panic!("Not a ZST")
         }
@@ -217,9 +219,9 @@ pub(crate) struct RefragmentHooks {
     hooks: HashMap<TypeId, fn(&mut World, Entity)>,
 }
 
-/// Command to set relationships between entities. If either of the participants do not exist or the
-/// host tries to target itself the operation will be ignored and logged.
-pub struct Set<R>
+/// Command to set relationship target for entities. If either of the participants do not exist or
+/// the host tries to target itself the operation will be ignored and logged.
+pub struct Target<R>
 where
     R: Relation,
 {
@@ -228,7 +230,7 @@ where
     pub _phantom: PhantomData<R>,
 }
 
-impl<R> Command for Set<R>
+impl<R> Command for Target<R>
 where
     R: Relation,
 {
@@ -238,35 +240,35 @@ where
 
         if self.host == self.target {
             warn!(
-                "Tried to set {} from {:?} to itself. \
+                "{host:?} Tried to target to itself with {rel}. \
                 Self referential relations are not allowed. \
                 Ignoring.",
-                std::any::type_name::<R>(),
-                self.host,
+                host = self.host,
+                rel = std::any::type_name::<R>(),
             );
             return;
         }
 
         if world.get_entity(self.target).is_none() {
             warn!(
-                "Tried to set {rel} from {from:?} to {to:?}. \
-                {to:?} does not exist. \
+                "{host:?} tried to target {target:?} with {rel}. \
+                {target:?} does not exist. \
                 Ignoring.",
+                host = self.host,
+                target = self.target,
                 rel = std::any::type_name::<R>(),
-                from = self.host,
-                to = self.target,
             );
             return;
         }
 
         if world.get_entity(self.host).is_none() {
             warn!(
-                "Tried to set {rel} from {from:?} to {to:?}. \
-                {from:?} does not exist. \
+                "{host:?} tried to target {target:?} with {rel}. \
+                {host:?} does not exist. \
                 Ignoring.",
+                host = self.host,
+                target = self.target,
                 rel = std::any::type_name::<R>(),
-                from = self.host,
-                to = self.target,
             );
             return;
         }
@@ -320,7 +322,7 @@ where
 
         if let Some(old) = old.filter(|old| R::EXCLUSIVE && self.target != *old) {
             Command::write(
-                Unset::<R> {
+                Untarget::<R> {
                     host: self.host,
                     target: old,
                     _phantom: PhantomData,
@@ -334,7 +336,7 @@ where
 /// Command to remove relationships between entities
 /// This operation is not noisy so if either participant does not exist or
 /// the relation does not exist nothing happens.
-pub struct Unset<R>
+pub struct Untarget<R>
 where
     R: Relation,
 {
@@ -343,13 +345,13 @@ where
     pub _phantom: PhantomData<R>,
 }
 
-impl<R: Relation> Command for Unset<R> {
+impl<R: Relation> Command for Untarget<R> {
     #[allow(clippy::let_unit_value)]
     fn write(self, world: &mut World) {
         let _ = R::ZST_OR_PANIC;
 
         Command::write(
-            UnsetErased {
+            UntargetErased {
                 host: self.host,
                 target: self.target,
                 typeid: TypeId::of::<R>(),
@@ -360,14 +362,14 @@ impl<R: Relation> Command for Unset<R> {
     }
 }
 
-struct UnsetErased {
+struct UntargetErased {
     host: Entity,
     target: Entity,
     typeid: TypeId,
     policy: CleanupPolicy,
 }
 
-impl Command for UnsetErased {
+impl Command for UntargetErased {
     fn write(self, world: &mut World) {
         let Some(refragment) = world
             .resource::<RefragmentHooks>()
@@ -447,8 +449,8 @@ impl Command for UnsetErased {
     }
 }
 
-/// Command for entities to to unset all targets of a given relation.
-pub struct UnsetAll<R>
+/// Command for entities to untarget all of their relations of a given type.
+pub struct UntargetAll<R>
 where
     R: Relation,
 {
@@ -456,7 +458,7 @@ where
     pub _phantom: PhantomData<R>,
 }
 
-impl<R: Relation> Command for UnsetAll<R> {
+impl<R: Relation> Command for UntargetAll<R> {
     #[allow(clippy::let_unit_value)]
     fn write(self, world: &mut World) {
         while let Some(target) = world
@@ -468,7 +470,7 @@ impl<R: Relation> Command for UnsetAll<R> {
             let _ = R::ZST_OR_PANIC;
 
             Command::write(
-                Unset::<R> {
+                Untarget::<R> {
                     target,
                     host: self.entity,
                     _phantom: PhantomData,
@@ -479,7 +481,7 @@ impl<R: Relation> Command for UnsetAll<R> {
     }
 }
 
-/// Command for entities to unset themselves from all relations of a given type.
+/// Command for entities to remove themselves as the target of all relations of a given type.
 pub struct Withdraw<R>
 where
     R: Relation,
@@ -500,7 +502,7 @@ impl<R: Relation> Command for Withdraw<R> {
             let _ = R::ZST_OR_PANIC;
 
             Command::write(
-                Unset::<R> {
+                Untarget::<R> {
                     host,
                     target: self.entity,
                     _phantom: PhantomData,
@@ -706,622 +708,5 @@ impl Command for CheckedDespawn {
                 refrag(world, entity);
             }
         }
-    }
-}
-
-/// Convenience trait to sugar using relation commands. Can be used with `World` or `Commands`.
-/// ```
-/// use bevy::prelude::*;
-/// use aery::prelude::*;
-///
-/// #[derive(Relation)]
-/// struct R;
-///
-/// fn sys(mut commands: Commands) {
-///     let a = commands.spawn_empty().id();
-///     let b = commands.spawn_empty().id();
-///     commands.set::<R>(a, b);
-/// }
-/// ```
-pub trait RelationCommands {
-    fn set<R: Relation>(&mut self, host: Entity, target: Entity);
-    fn unset<R: Relation>(&mut self, host: Entity, target: Entity);
-    fn unset_all<R: Relation>(&mut self, entity: Entity);
-    fn withdraw<R: Relation>(&mut self, entity: Entity);
-    fn checked_despawn(&mut self, entity: Entity);
-}
-
-impl RelationCommands for Commands<'_, '_> {
-    fn set<R: Relation>(&mut self, host: Entity, target: Entity) {
-        self.add(Set::<R> {
-            host,
-            target,
-            _phantom: PhantomData,
-        });
-    }
-
-    fn unset<R: Relation>(&mut self, host: Entity, target: Entity) {
-        self.add(Unset::<R> {
-            host,
-            target,
-            _phantom: PhantomData,
-        });
-    }
-
-    fn unset_all<R: Relation>(&mut self, entity: Entity) {
-        self.add(UnsetAll::<R> {
-            entity,
-            _phantom: PhantomData,
-        });
-    }
-
-    fn withdraw<R: Relation>(&mut self, entity: Entity) {
-        self.add(Withdraw::<R> {
-            entity,
-            _phantom: PhantomData,
-        });
-    }
-
-    fn checked_despawn(&mut self, entity: Entity) {
-        self.add(CheckedDespawn { entity });
-    }
-}
-
-impl RelationCommands for World {
-    fn set<R: Relation>(&mut self, host: Entity, target: Entity) {
-        Command::write(
-            Set::<R> {
-                host,
-                target,
-                _phantom: PhantomData,
-            },
-            self,
-        );
-    }
-
-    fn unset<R: Relation>(&mut self, host: Entity, target: Entity) {
-        Command::write(
-            Unset::<R> {
-                host,
-                target,
-                _phantom: PhantomData,
-            },
-            self,
-        );
-    }
-
-    fn unset_all<R: Relation>(&mut self, entity: Entity) {
-        Command::write(
-            UnsetAll::<R> {
-                entity,
-                _phantom: PhantomData,
-            },
-            self,
-        );
-    }
-
-    fn withdraw<R: Relation>(&mut self, entity: Entity) {
-        Command::write(
-            Withdraw::<R> {
-                entity,
-                _phantom: PhantomData,
-            },
-            self,
-        );
-    }
-
-    fn checked_despawn(&mut self, entity: Entity) {
-        Command::write(CheckedDespawn { entity }, self);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        prelude::*,
-        relation::{Edges, Participant, RefragmentHooks, RootMarker},
-    };
-    use bevy::prelude::*;
-    use core::any::TypeId;
-    use std::array::from_fn;
-
-    fn has_edges(world: &World, entity: Entity) -> bool {
-        world.get::<Edges>(entity).is_some()
-    }
-
-    fn is_root<R: Relation>(world: &World, entity: Entity) -> bool {
-        world.get::<RootMarker<R>>(entity).is_some()
-    }
-
-    fn is_participant<R: Relation>(world: &World, entity: Entity) -> bool {
-        world.get::<Participant<R>>(entity).is_some()
-    }
-
-    fn targeting<R: Relation>(world: &World, host: Entity, target: Entity) -> bool {
-        let host_is_targeting = world
-            .get::<Edges>(host)
-            .map(|edges| &edges.targets[R::CLEANUP_POLICY as usize])
-            .and_then(|bucket| bucket.get(&TypeId::of::<R>()))
-            .map_or(false, |set| set.contains(&target));
-
-        let target_is_hosted = world
-            .get::<Edges>(target)
-            .map(|edges| &edges.hosts[R::CLEANUP_POLICY as usize])
-            .and_then(|bucket| bucket.get(&TypeId::of::<R>()))
-            .map_or(false, |set| set.contains(&host));
-
-        if host_is_targeting != target_is_hosted {
-            panic!("Asymmetric edge info");
-        }
-
-        host_is_targeting
-    }
-
-    #[test]
-    fn set_unset() {
-        #[derive(Relation)]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-        let [host, target] = from_fn(|_| world.spawn_empty().id());
-
-        world.set::<R>(host, target);
-        assert!(targeting::<R>(&world, host, target));
-        assert!(is_participant::<R>(&world, host));
-        assert!(is_root::<R>(&world, target));
-
-        world.unset::<R>(host, target);
-        assert!(!has_edges(&world, target));
-        assert!(!has_edges(&world, host));
-        assert!(!is_participant::<R>(&world, host));
-        assert!(!is_root::<R>(&world, target));
-    }
-
-    #[test]
-    fn exclusive() {
-        #[derive(Relation)]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-        let [host, t0, t1] = from_fn(|_| world.spawn_empty().id());
-
-        // Before overwrite
-        world.set::<R>(host, t0);
-
-        assert!(targeting::<R>(&world, host, t0));
-        assert!(is_participant::<R>(&world, host));
-        assert!(is_root::<R>(&world, t0));
-
-        // After overwrite
-        world.set::<R>(host, t1);
-
-        assert!(targeting::<R>(&world, host, t1));
-        assert!(is_participant::<R>(&world, host));
-        assert!(is_root::<R>(&world, t1));
-
-        assert!(!has_edges(&world, t0));
-        assert!(!is_root::<R>(&world, t0));
-    }
-
-    #[derive(Relation)]
-    #[cleanup(policy = "Orphan")]
-    struct Orphan;
-
-    #[derive(Relation)]
-    #[cleanup(policy = "Counted")]
-    struct Counted;
-
-    #[derive(Relation)]
-    #[cleanup(policy = "Recursive")]
-    struct Recursive;
-
-    #[derive(Relation)]
-    #[cleanup(policy = "Total")]
-    struct Total;
-
-    #[derive(Debug)]
-    struct TestEdges {
-        orphan: Entity,
-        counted: Entity,
-        recursive: Entity,
-        total: Entity,
-    }
-
-    #[derive(Debug)]
-    struct Test {
-        center: Entity,
-        targets: TestEdges,
-        hosts: TestEdges,
-    }
-
-    impl Test {
-        fn new(world: &mut World) -> Self {
-            let test = Self {
-                center: world.spawn_empty().id(),
-                targets: TestEdges {
-                    orphan: world.spawn_empty().id(),
-                    counted: world.spawn_empty().id(),
-                    recursive: world.spawn_empty().id(),
-                    total: world.spawn_empty().id(),
-                },
-                hosts: TestEdges {
-                    orphan: world.spawn_empty().id(),
-                    counted: world.spawn_empty().id(),
-                    recursive: world.spawn_empty().id(),
-                    total: world.spawn_empty().id(),
-                },
-            };
-
-            world.set::<Orphan>(test.hosts.orphan, test.center);
-            world.set::<Orphan>(test.center, test.targets.orphan);
-
-            world.set::<Counted>(test.hosts.counted, test.center);
-            world.set::<Counted>(test.center, test.targets.counted);
-
-            world.set::<Recursive>(test.hosts.recursive, test.center);
-            world.set::<Recursive>(test.center, test.targets.recursive);
-
-            world.set::<Total>(test.hosts.total, test.center);
-            world.set::<Total>(test.center, test.targets.total);
-
-            test
-        }
-
-        fn assert_unchanged(&self, world: &World) {
-            assert!(targeting::<Orphan>(world, self.hosts.orphan, self.center));
-            assert!(targeting::<Orphan>(world, self.center, self.targets.orphan));
-            assert!(is_participant::<Orphan>(world, self.hosts.orphan,));
-            assert!(is_root::<Orphan>(world, self.targets.orphan));
-
-            assert!(targeting::<Counted>(world, self.hosts.counted, self.center));
-            assert!(targeting::<Counted>(
-                world,
-                self.center,
-                self.targets.counted
-            ));
-            assert!(is_participant::<Counted>(world, self.hosts.counted,));
-            assert!(is_root::<Counted>(world, self.targets.counted));
-
-            assert!(targeting::<Recursive>(
-                world,
-                self.hosts.recursive,
-                self.center
-            ));
-            assert!(targeting::<Recursive>(
-                world,
-                self.center,
-                self.targets.recursive
-            ));
-            assert!(is_participant::<Recursive>(world, self.hosts.recursive,));
-            assert!(is_root::<Recursive>(world, self.targets.recursive));
-
-            assert!(targeting::<Total>(world, self.hosts.total, self.center));
-            assert!(targeting::<Total>(world, self.center, self.targets.total));
-            assert!(is_participant::<Total>(world, self.hosts.total));
-            assert!(is_root::<Total>(world, self.targets.total));
-
-            assert!(is_participant::<Orphan>(world, self.center));
-            assert!(is_participant::<Counted>(world, self.center));
-            assert!(is_participant::<Recursive>(world, self.center));
-            assert!(is_participant::<Total>(world, self.center));
-        }
-
-        fn assert_cleaned(&self, world: &World) {
-            assert!(world.get_entity(self.center).is_none());
-
-            assert!(!has_edges(world, self.hosts.orphan));
-            assert!(!has_edges(world, self.targets.orphan));
-            assert!(!is_participant::<Orphan>(world, self.hosts.orphan));
-            assert!(!is_root::<Orphan>(world, self.targets.orphan));
-
-            assert!(world.get_entity(self.targets.counted).is_none());
-            assert!(!has_edges(world, self.hosts.counted));
-            assert!(!is_participant::<Counted>(world, self.hosts.counted,));
-
-            assert!(world.get_entity(self.hosts.recursive).is_none());
-            assert!(!has_edges(world, self.targets.recursive));
-            assert!(!is_root::<Recursive>(world, self.targets.recursive));
-
-            assert!(world.get_entity(self.hosts.total).is_none());
-            assert!(world.get_entity(self.targets.total).is_none());
-        }
-    }
-
-    #[test]
-    fn orphan_in_despawned() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Orphan")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(e, test.center);
-
-        world.checked_despawn(e);
-        test.assert_unchanged(&world);
-        assert!(!is_participant::<R>(&world, test.center));
-    }
-
-    #[test]
-    fn orphan_out_despawned() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Orphan")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(test.center, e);
-
-        world.checked_despawn(e);
-        test.assert_unchanged(&world);
-        assert!(!is_participant::<R>(&world, test.center));
-    }
-
-    #[test]
-    fn counted_in_despawned() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Counted")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(e, test.center);
-
-        world.checked_despawn(e);
-        test.assert_cleaned(&world);
-    }
-
-    #[test]
-    fn counted_out_despawned() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Counted")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(test.center, e);
-
-        world.checked_despawn(e);
-        test.assert_unchanged(&world);
-        assert!(!is_participant::<R>(&world, test.center));
-    }
-
-    #[test]
-    fn recursive_in_despawned() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Recursive")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(e, test.center);
-
-        world.checked_despawn(e);
-        test.assert_unchanged(&world);
-        assert!(!is_participant::<R>(&world, test.center));
-    }
-
-    #[test]
-    fn recursive_out_despawned() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Recursive")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(test.center, e);
-
-        world.checked_despawn(e);
-        test.assert_cleaned(&world);
-    }
-
-    #[test]
-    fn total_in_despawned() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Total")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(e, test.center);
-
-        world.checked_despawn(e);
-        test.assert_cleaned(&world);
-    }
-
-    #[test]
-    fn total_out_despawned() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Total")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(test.center, e);
-
-        world.checked_despawn(e);
-        test.assert_cleaned(&world);
-    }
-
-    #[test]
-    fn orphan_in_unset() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Orphan")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(e, test.center);
-        world.unset::<R>(e, test.center);
-
-        test.assert_unchanged(&world);
-        assert!(!is_participant::<R>(&world, test.center));
-    }
-
-    #[test]
-    fn orphan_out_unset() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Orphan")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(test.center, e);
-        world.unset::<R>(test.center, e);
-
-        test.assert_unchanged(&world);
-        assert!(!is_participant::<R>(&world, test.center));
-    }
-
-    #[test]
-    fn counted_in_unset() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Counted")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(e, test.center);
-        world.unset::<R>(e, test.center);
-
-        test.assert_cleaned(&world);
-    }
-
-    #[test]
-    fn counted_out_unset() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Counted")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(test.center, e);
-        world.unset::<R>(test.center, e);
-
-        test.assert_unchanged(&world);
-        assert!(!is_participant::<R>(&world, test.center));
-    }
-
-    #[test]
-    fn recursive_in_unset() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Recursive")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(e, test.center);
-        world.unset::<R>(e, test.center);
-
-        test.assert_unchanged(&world);
-        assert!(!is_participant::<R>(&world, test.center));
-    }
-
-    #[test]
-    fn recursive_out_unset() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Recursive")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(test.center, e);
-        world.unset::<R>(test.center, e);
-
-        test.assert_cleaned(&world);
-    }
-
-    #[test]
-    fn total_in_unset() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Total")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(e, test.center);
-        world.unset::<R>(e, test.center);
-
-        test.assert_cleaned(&world);
-    }
-
-    #[test]
-    fn total_out_unset() {
-        #[derive(Relation)]
-        #[cleanup(policy = "Total")]
-        struct R;
-
-        let mut world = World::new();
-        world.init_resource::<RefragmentHooks>();
-
-        let test = Test::new(&mut world);
-
-        let e = world.spawn_empty().id();
-        world.set::<R>(test.center, e);
-        world.unset::<R>(test.center, e);
-
-        test.assert_cleaned(&world);
     }
 }
